+++ lib/boruta/openid.ex
@@ -1,6 +1,7 @@
 defmodule Boruta.OpenidModule do
   @moduledoc false
   @callback jwks(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @callback userinfo(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
 end
 
 defmodule Boruta.Openid do
@@ -9,14 +10,29 @@ defmodule Boruta.Openid do
 
   > __Note__: this module follows inverted heaxagonal architecture, its functions will invoke callbacks of the given module argument and return its result.
   >
-  > The definition of those callbacks are provided by either `Boruta.Openid.Application` or `Boruta.Oauth.JwksApplication`
+  > The definition of those callbacks are provided by either `Boruta.Openid.Application` or `Boruta.Openid.JwksApplication` and `Boruta.Openid.UserinfoApplication`
   """
 
   import Boruta.Config, only: [clients: 0]
+
+  alias Boruta.Oauth.Authorization.AccessToken
+  alias Boruta.Oauth.BearerToken
+  alias Boruta.Oauth.Token
 
   def jwks(conn, module) do
     jwk_keys = clients().list_clients_jwk()
 
     module.jwk_list(conn, jwk_keys)
+  end
+
+  def userinfo(conn, module) do
+    with {:ok, access_token} <- BearerToken.extract_token(conn),
+         {:ok, token} <- AccessToken.authorize(value: access_token),
+         {:ok, userinfo} <- Token.userinfo(token) do
+      module.userinfo_fetched(conn, userinfo)
+    else
+      {:error, error} ->
+        module.unauthorized(conn, error)
+    end
   end
 end