+++ test/support/boruta/openid/application_mock.ex
@@ -4,4 +4,10 @@ defmodule Boruta.Openid.ApplicationMock do
 
   @impl Boruta.Openid.Application
   def jwk_list(_conn, jwk_keys), do: {:jwk_list, jwk_keys}
+
+  @impl Boruta.Openid.Application
+  def userinfo_fetched(_conn, userinfo), do: {:userinfo_fetched, userinfo}
+
+  @impl Boruta.Openid.Application
+  def unauthorized(_conn, error), do: {:unauthorized, error}
 end